requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lightning-utilities==0.8.0
lit==16.0.5
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
psutil==5.9.5
pyparsing==3.0.9
pytorch-lightning==2.0.2
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torch-geometric==2.3.1
torchmetrics==0.11.4
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.0
urllib3==2.0.2
yarl==1.9.2
